refacor: Entity에 null 들어올 수 있으므로 Long으로 변경

+++ src/main/java/dazaram/eureka/elastic/domain/RecipeSequenceDocument.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class RecipeSequenceDocument {
 	private String content;
-	private int sequence;
+	private Long sequence;
 
 	public RecipeSequenceDocument(RecipeSequence recipeSequence) {
 		this.content = recipeSequence.getContent();
